Generics for matrix

+++ src/test/java/com/janoz/aoc/y2021/day6/Day6Test.java
@@ -3,8 +3,6 @@ package com.janoz.aoc.y2021.day6;
 
 import org.junit.jupiter.api.Test;
 
-import java.math.BigInteger;
-
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.CoreMatchers.equalTo;
 
@@ -13,11 +11,11 @@ class Day6Test {
 
     @Test
     void testPart1() {
-        assertThat(new Day6().calculate("3,4,3,1,2", 80), equalTo(BigInteger.valueOf(5934L)));
+        assertThat(new Day6().calculate("3,4,3,1,2", 80), equalTo(5934L));
     }
 
     @Test
     void testPart2() {
-        assertThat(new Day6().calculate("3,4,3,1,2", 256), equalTo(BigInteger.valueOf(26984457539L)));
+        assertThat(new Day6().calculate("3,4,3,1,2", 256), equalTo(26984457539L));
     }
 }
